requirements.txt
asgiref==3.3.1
certifi==2020.12.5
chardet==4.0.0
colorama==0.4.4
colorful==0.5.4
Django==3.1.5
django-crispy-forms==1.10.0
django-filter==2.4.0
djangorestframework==3.12.2
idna==2.10
mysqlclient==2.0.3
Pillow==8.1.0
prettyprinter==0.18.0
Pygments==2.7.3
pytz==2020.5
requests==2.25.1
sqlparse==0.4.1
urllib3==1.26.2
